Move Tee settings to `menus_settings_tee.cpp`
