+++ src/game/client/components/menus_settings_tee.cpp
@@ -1,0 +1,504 @@
+/* (c) Magnus Auvinen. See licence.txt in the root of the distribution for more information. */
+/* If you are missing that file, acquire a complete release at teeworlds.com.                */
+#include "menus.h"
+
+#include <base/str.h>
+
+#include <engine/font_icons.h>
+#include <engine/graphics.h>
+#include <engine/shared/config.h>
+#include <engine/storage.h>
+#include <engine/textrender.h>
+
+#include <game/client/animstate.h>
+#include <game/client/components/console.h>
+#include <game/client/components/emoticon.h>
+#include <game/client/components/skins.h>
+#include <game/client/components/tooltips.h>
+#include <game/client/gameclient.h>
+#include <game/client/skin.h>
+#include <game/client/ui.h>
+#include <game/client/ui_listbox.h>
+#include <game/localization.h>
+
+#include <algorithm>
+#include <vector>
+
+void CMenus::RenderSettingsTee(CUIRect MainView)
+{
+	CUIRect TabBar, PlayerTab, DummyTab, ChangeInfo;
+	MainView.HSplitTop(20.0f, &TabBar, &MainView);
+	TabBar.VSplitMid(&TabBar, &ChangeInfo, 20.f);
+	TabBar.VSplitMid(&PlayerTab, &DummyTab);
+	MainView.HSplitTop(10.0f, nullptr, &MainView);
+
+	static CButtonContainer s_PlayerTabButton;
+	if(DoButton_MenuTab(&s_PlayerTabButton, Localize("Player"), !m_Dummy, &PlayerTab, IGraphics::CORNER_L, nullptr, nullptr, nullptr, nullptr, 4.0f))
+	{
+		m_Dummy = false;
+		m_SkinListScrollToSelected = true;
+	}
+
+	static CButtonContainer s_DummyTabButton;
+	if(DoButton_MenuTab(&s_DummyTabButton, Localize("Dummy"), m_Dummy, &DummyTab, IGraphics::CORNER_R, nullptr, nullptr, nullptr, nullptr, 4.0f))
+	{
+		m_Dummy = true;
+		m_SkinListScrollToSelected = true;
+	}
+
+	if(Client()->State() == IClient::STATE_ONLINE &&
+		GameClient()->m_aNextChangeInfo[m_Dummy] > Client()->GameTick(m_Dummy))
+	{
+		char aChangeInfo[128], aTimeLeft[32];
+		str_format(aTimeLeft, sizeof(aTimeLeft), Localize("%ds left"), (GameClient()->m_aNextChangeInfo[m_Dummy] - Client()->GameTick(m_Dummy) + Client()->GameTickSpeed() - 1) / Client()->GameTickSpeed());
+		str_format(aChangeInfo, sizeof(aChangeInfo), "%s: %s", Localize("Player info change cooldown"), aTimeLeft);
+		Ui()->DoLabel(&ChangeInfo, aChangeInfo, 10.f, TEXTALIGN_ML);
+	}
+
+	if(g_Config.m_Debug)
+	{
+		const CSkins::CSkinLoadingStats Stats = GameClient()->m_Skins.LoadingStats();
+		char aStats[256];
+		str_format(aStats, sizeof(aStats), "unloaded: %" PRIzu ", pending: %" PRIzu ", loading: %" PRIzu ",\nloaded: %" PRIzu ", error: %" PRIzu ", notfound: %" PRIzu,
+			Stats.m_NumUnloaded, Stats.m_NumPending, Stats.m_NumLoading, Stats.m_NumLoaded, Stats.m_NumError, Stats.m_NumNotFound);
+		Ui()->DoLabel(&ChangeInfo, aStats, 9.0f, TEXTALIGN_MR);
+	}
+
+	char *pSkinName;
+	size_t SkinNameSize;
+	int *pUseCustomColor;
+	unsigned *pColorBody;
+	unsigned *pColorFeet;
+	int *pEmote;
+	if(!m_Dummy)
+	{
+		pSkinName = g_Config.m_ClPlayerSkin;
+		SkinNameSize = sizeof(g_Config.m_ClPlayerSkin);
+		pUseCustomColor = &g_Config.m_ClPlayerUseCustomColor;
+		pColorBody = &g_Config.m_ClPlayerColorBody;
+		pColorFeet = &g_Config.m_ClPlayerColorFeet;
+		pEmote = &g_Config.m_ClPlayerDefaultEyes;
+	}
+	else
+	{
+		pSkinName = g_Config.m_ClDummySkin;
+		SkinNameSize = sizeof(g_Config.m_ClDummySkin);
+		pUseCustomColor = &g_Config.m_ClDummyUseCustomColor;
+		pColorBody = &g_Config.m_ClDummyColorBody;
+		pColorFeet = &g_Config.m_ClDummyColorFeet;
+		pEmote = &g_Config.m_ClDummyDefaultEyes;
+	}
+
+	const float EyeButtonSize = 40.0f;
+	const bool RenderEyesBelow = MainView.w < 750.0f;
+	CUIRect YourSkin, Checkboxes, SkinPrefix, Eyes, Button, Label;
+	MainView.HSplitTop(90.0f, &YourSkin, &MainView);
+	if(RenderEyesBelow)
+	{
+		YourSkin.VSplitLeft(MainView.w * 0.45f, &YourSkin, &Checkboxes);
+		Checkboxes.VSplitLeft(MainView.w * 0.35f, &Checkboxes, &SkinPrefix);
+		MainView.HSplitTop(5.0f, nullptr, &MainView);
+		MainView.HSplitTop(EyeButtonSize, &Eyes, &MainView);
+		Eyes.VSplitRight(EyeButtonSize * (float)NUM_EMOTES + 5.0f * (float)(NUM_EMOTES - 1), nullptr, &Eyes);
+	}
+	else
+	{
+		YourSkin.VSplitRight(3 * EyeButtonSize + 2 * 5.0f, &YourSkin, &Eyes);
+		const float RemainderWidth = YourSkin.w;
+		YourSkin.VSplitLeft(RemainderWidth * 0.4f, &YourSkin, &Checkboxes);
+		Checkboxes.VSplitLeft(RemainderWidth * 0.35f, &Checkboxes, &SkinPrefix);
+		SkinPrefix.VSplitRight(20.0f, &SkinPrefix, nullptr);
+	}
+	YourSkin.VSplitRight(20.0f, &YourSkin, nullptr);
+	Checkboxes.VSplitRight(20.0f, &Checkboxes, nullptr);
+
+	// Checkboxes
+	bool ShouldRefresh = false;
+	Checkboxes.HSplitTop(20.0f, &Button, &Checkboxes);
+	if(DoButton_CheckBox(&g_Config.m_ClDownloadSkins, Localize("Download skins"), g_Config.m_ClDownloadSkins, &Button))
+	{
+		g_Config.m_ClDownloadSkins ^= 1;
+		ShouldRefresh = true;
+	}
+
+	Checkboxes.HSplitTop(20.0f, &Button, &Checkboxes);
+	if(DoButton_CheckBox(&g_Config.m_ClDownloadCommunitySkins, Localize("Download community skins"), g_Config.m_ClDownloadCommunitySkins, &Button))
+	{
+		g_Config.m_ClDownloadCommunitySkins ^= 1;
+		ShouldRefresh = true;
+	}
+
+	Checkboxes.HSplitTop(20.0f, &Button, &Checkboxes);
+	if(DoButton_CheckBox(&g_Config.m_ClVanillaSkinsOnly, Localize("Vanilla skins only"), g_Config.m_ClVanillaSkinsOnly, &Button))
+	{
+		g_Config.m_ClVanillaSkinsOnly ^= 1;
+		ShouldRefresh = true;
+	}
+
+	Checkboxes.HSplitTop(20.0f, &Button, &Checkboxes);
+	if(DoButton_CheckBox(&g_Config.m_ClFatSkins, Localize("Fat skins (DDFat)"), g_Config.m_ClFatSkins, &Button))
+	{
+		g_Config.m_ClFatSkins ^= 1;
+	}
+
+	// Skin prefix
+	{
+		SkinPrefix.HSplitTop(20.0f, &Label, &SkinPrefix);
+		Ui()->DoLabel(&Label, Localize("Skin prefix"), 14.0f, TEXTALIGN_ML);
+
+		SkinPrefix.HSplitTop(20.0f, &Button, &SkinPrefix);
+		static CLineInput s_SkinPrefixInput(g_Config.m_ClSkinPrefix, sizeof(g_Config.m_ClSkinPrefix));
+		if(Ui()->DoClearableEditBox(&s_SkinPrefixInput, &Button, 14.0f))
+		{
+			ShouldRefresh = true;
+		}
+
+		SkinPrefix.HSplitTop(2.0f, nullptr, &SkinPrefix);
+
+		static const char *s_apSkinPrefixes[] = {"kitty", "santa"};
+		static CButtonContainer s_aPrefixButtons[std::size(s_apSkinPrefixes)];
+		for(size_t i = 0; i < std::size(s_apSkinPrefixes); i++)
+		{
+			SkinPrefix.HSplitTop(20.0f, &Button, &SkinPrefix);
+			Button.HMargin(2.0f, &Button);
+			if(DoButton_Menu(&s_aPrefixButtons[i], s_apSkinPrefixes[i], 0, &Button))
+			{
+				str_copy(g_Config.m_ClSkinPrefix, s_apSkinPrefixes[i]);
+				ShouldRefresh = true;
+			}
+		}
+	}
+
+	// Player skin area
+	CUIRect CustomColorsButton, RandomSkinButton;
+	YourSkin.HSplitTop(20.0f, &Label, &YourSkin);
+	YourSkin.HSplitBottom(20.0f, &YourSkin, &CustomColorsButton);
+	CustomColorsButton.VSplitRight(30.0f, &CustomColorsButton, &RandomSkinButton);
+	CustomColorsButton.VSplitRight(20.0f, &CustomColorsButton, nullptr);
+	YourSkin.VSplitLeft(65.0f, &YourSkin, &Button);
+	Button.VSplitLeft(5.0f, nullptr, &Button);
+	Button.HMargin((Button.h - 20.0f) / 2.0f, &Button);
+
+	char aBuf[128 + IO_MAX_PATH_LENGTH];
+	str_format(aBuf, sizeof(aBuf), "%s:", Localize("Your skin"));
+	Ui()->DoLabel(&Label, aBuf, 14.0f, TEXTALIGN_ML);
+
+	CSkins::CSkinList &SkinList = GameClient()->m_Skins.SkinList();
+	const CSkin *pDefaultSkin = GameClient()->m_Skins.Find("default");
+	const CSkins::CSkinContainer *pOwnSkinContainer = GameClient()->m_Skins.FindContainerOrNullptr(pSkinName[0] == '\0' ? "default" : pSkinName);
+	if(pOwnSkinContainer != nullptr && pOwnSkinContainer->IsSpecial())
+	{
+		pOwnSkinContainer = nullptr; // Special skins cannot be selected, show as missing due to invalid name
+	}
+
+	CTeeRenderInfo OwnSkinInfo;
+	OwnSkinInfo.Apply(pOwnSkinContainer == nullptr || pOwnSkinContainer->Skin() == nullptr ? pDefaultSkin : pOwnSkinContainer->Skin().get());
+	OwnSkinInfo.ApplyColors(*pUseCustomColor, *pColorBody, *pColorFeet);
+	OwnSkinInfo.m_Size = 50.0f;
+
+	// Tee
+	{
+		vec2 OffsetToMid;
+		CRenderTools::GetRenderTeeOffsetToRenderedTee(CAnimState::GetIdle(), &OwnSkinInfo, OffsetToMid);
+		const vec2 TeeRenderPos = vec2(YourSkin.x + YourSkin.w / 2.0f, YourSkin.y + YourSkin.h / 2.0f + OffsetToMid.y);
+		// tee looking towards cursor, and it is happy when you touch it
+		const vec2 DeltaPosition = Ui()->MousePos() - TeeRenderPos;
+		const float Distance = length(DeltaPosition);
+		const float InteractionDistance = 20.0f;
+		const vec2 TeeDirection = Distance < InteractionDistance ? normalize(vec2(DeltaPosition.x, std::max(DeltaPosition.y, 0.5f))) : normalize(DeltaPosition);
+		const int TeeEmote = Distance < InteractionDistance ? EMOTE_HAPPY : *pEmote;
+		RenderTools()->RenderTee(CAnimState::GetIdle(), &OwnSkinInfo, TeeEmote, TeeDirection, TeeRenderPos);
+	}
+
+	// Skin loading status
+	const auto &&RenderSkinStatus = [&](CUIRect Parent, const CSkins::CSkinContainer *pSkinContainer, const void *pStatusTooltipId) {
+		if(pSkinContainer != nullptr && pSkinContainer->State() == CSkins::CSkinContainer::EState::LOADED)
+		{
+			return;
+		}
+
+		CUIRect StatusIcon;
+		Parent.HSplitTop(20.0f, &StatusIcon, nullptr);
+		StatusIcon.VSplitLeft(20.0f, &StatusIcon, nullptr);
+
+		if(pSkinContainer != nullptr &&
+			(pSkinContainer->State() == CSkins::CSkinContainer::EState::UNLOADED ||
+				pSkinContainer->State() == CSkins::CSkinContainer::EState::PENDING ||
+				pSkinContainer->State() == CSkins::CSkinContainer::EState::LOADING))
+		{
+			Ui()->RenderProgressSpinner(StatusIcon.Center(), 5.0f);
+		}
+		else
+		{
+			TextRender()->TextColor(ColorRGBA(1.0f, 0.0f, 0.0f, 1.0f));
+			TextRender()->SetFontPreset(EFontPreset::ICON_FONT);
+			TextRender()->SetRenderFlags(ETextRenderFlags::TEXT_RENDER_FLAG_ONLY_ADVANCE_WIDTH | ETextRenderFlags::TEXT_RENDER_FLAG_NO_X_BEARING | ETextRenderFlags::TEXT_RENDER_FLAG_NO_Y_BEARING | ETextRenderFlags::TEXT_RENDER_FLAG_NO_PIXEL_ALIGNMENT | ETextRenderFlags::TEXT_RENDER_FLAG_NO_OVERSIZE);
+			Ui()->DoLabel(&StatusIcon, pSkinContainer == nullptr || pSkinContainer->State() == CSkins::CSkinContainer::EState::ERROR ? FontIcon::TRIANGLE_EXCLAMATION : FontIcon::QUESTION, 12.0f, TEXTALIGN_MC);
+			TextRender()->SetRenderFlags(0);
+			TextRender()->SetFontPreset(EFontPreset::DEFAULT_FONT);
+			TextRender()->TextColor(TextRender()->DefaultTextColor());
+			Ui()->DoButtonLogic(pStatusTooltipId, 0, &StatusIcon, BUTTONFLAG_NONE);
+			const char *pErrorTooltip;
+			if(pSkinContainer == nullptr)
+			{
+				pErrorTooltip = Localize("This skin name cannot be used.");
+			}
+			else if(pSkinContainer->State() == CSkins::CSkinContainer::EState::ERROR)
+			{
+				pErrorTooltip = Localize("Skin could not be loaded due to an error. Check the local console for details.");
+			}
+			else
+			{
+				pErrorTooltip = Localize("Skin could not be found.");
+			}
+			GameClient()->m_Tooltips.DoToolTip(pStatusTooltipId, &StatusIcon, pErrorTooltip);
+		}
+	};
+	static char s_StatusTooltipId;
+	RenderSkinStatus(YourSkin, pOwnSkinContainer, &s_StatusTooltipId);
+
+	// Skin name
+	static CLineInput s_SkinInput;
+	s_SkinInput.SetBuffer(pSkinName, SkinNameSize);
+	s_SkinInput.SetEmptyText("default");
+	if(Ui()->DoClearableEditBox(&s_SkinInput, &Button, 14.0f))
+	{
+		SetNeedSendInfo();
+		m_SkinListScrollToSelected = true;
+		SkinList.ForceRefresh();
+	}
+
+	// Random skin button
+	static CButtonContainer s_RandomSkinButton;
+	static const char *s_apDice[] = {FontIcon::DICE_ONE, FontIcon::DICE_TWO, FontIcon::DICE_THREE, FontIcon::DICE_FOUR, FontIcon::DICE_FIVE, FontIcon::DICE_SIX};
+	static int s_CurrentDie = rand() % std::size(s_apDice);
+	TextRender()->SetFontPreset(EFontPreset::ICON_FONT);
+	TextRender()->SetRenderFlags(ETextRenderFlags::TEXT_RENDER_FLAG_ONLY_ADVANCE_WIDTH | ETextRenderFlags::TEXT_RENDER_FLAG_NO_X_BEARING | ETextRenderFlags::TEXT_RENDER_FLAG_NO_Y_BEARING | ETextRenderFlags::TEXT_RENDER_FLAG_NO_PIXEL_ALIGNMENT | ETextRenderFlags::TEXT_RENDER_FLAG_NO_OVERSIZE);
+	if(DoButton_Menu(&s_RandomSkinButton, s_apDice[s_CurrentDie], 0, &RandomSkinButton, BUTTONFLAG_LEFT, nullptr, IGraphics::CORNER_ALL, 5.0f, -0.2f))
+	{
+		GameClient()->m_Skins.RandomizeSkin(m_Dummy);
+		SetNeedSendInfo();
+		m_SkinListScrollToSelected = true;
+		s_CurrentDie = rand() % std::size(s_apDice);
+	}
+	TextRender()->SetRenderFlags(0);
+	TextRender()->SetFontPreset(EFontPreset::DEFAULT_FONT);
+	GameClient()->m_Tooltips.DoToolTip(&s_RandomSkinButton, &RandomSkinButton, Localize("Create a random skin"));
+
+	// Custom colors button
+	if(DoButton_CheckBox(pUseCustomColor, Localize("Custom colors"), *pUseCustomColor, &CustomColorsButton))
+	{
+		*pUseCustomColor = *pUseCustomColor ? 0 : 1;
+		SetNeedSendInfo();
+	}
+
+	// Default eyes
+	{
+		CTeeRenderInfo EyeSkinInfo = OwnSkinInfo;
+		EyeSkinInfo.m_Size = EyeButtonSize;
+		vec2 OffsetToMid;
+		CRenderTools::GetRenderTeeOffsetToRenderedTee(CAnimState::GetIdle(), &EyeSkinInfo, OffsetToMid);
+
+		CUIRect EyesRow;
+		Eyes.HSplitTop(EyeButtonSize, &EyesRow, &Eyes);
+		static CButtonContainer s_aEyeButtons[NUM_EMOTES];
+		for(int CurrentEyeEmote = 0; CurrentEyeEmote < NUM_EMOTES; CurrentEyeEmote++)
+		{
+			EyesRow.VSplitLeft(EyeButtonSize, &Button, &EyesRow);
+			EyesRow.VSplitLeft(5.0f, nullptr, &EyesRow);
+			if(!RenderEyesBelow && (CurrentEyeEmote + 1) % 3 == 0)
+			{
+				Eyes.HSplitTop(5.0f, nullptr, &Eyes);
+				Eyes.HSplitTop(EyeButtonSize, &EyesRow, &Eyes);
+			}
+
+			const ColorRGBA EyeButtonColor = ColorRGBA(1.0f, 1.0f, 1.0f, 0.25f + (*pEmote == CurrentEyeEmote ? 0.25f : 0.0f));
+			if(DoButton_Menu(&s_aEyeButtons[CurrentEyeEmote], "", 0, &Button, BUTTONFLAG_LEFT, nullptr, IGraphics::CORNER_ALL, 5.0f, 0.0f, EyeButtonColor))
+			{
+				*pEmote = CurrentEyeEmote;
+				if((int)m_Dummy == g_Config.m_ClDummy)
+					GameClient()->m_Emoticon.EyeEmote(CurrentEyeEmote);
+			}
+			GameClient()->m_Tooltips.DoToolTip(&s_aEyeButtons[CurrentEyeEmote], &Button, Localize("Choose default eyes when joining a server"));
+			RenderTools()->RenderTee(CAnimState::GetIdle(), &EyeSkinInfo, CurrentEyeEmote, vec2(1.0f, 0.0f), vec2(Button.x + Button.w / 2.0f, Button.y + Button.h / 2.0f + OffsetToMid.y));
+		}
+	}
+
+	// Custom color pickers
+	MainView.HSplitTop(5.0f, nullptr, &MainView);
+	if(*pUseCustomColor)
+	{
+		CUIRect CustomColors;
+		MainView.HSplitTop(95.0f, &CustomColors, &MainView);
+		CUIRect aRects[2];
+		CustomColors.VSplitMid(&aRects[0], &aRects[1], 20.0f);
+
+		unsigned *apColors[] = {pColorBody, pColorFeet};
+		const char *apParts[] = {Localize("Body"), Localize("Feet")};
+
+		for(int i = 0; i < 2; i++)
+		{
+			aRects[i].HSplitTop(20.0f, &Label, &aRects[i]);
+			Ui()->DoLabel(&Label, apParts[i], 14.0f, TEXTALIGN_ML);
+			if(RenderHslaScrollbars(&aRects[i], apColors[i], false, ColorHSLA::DARKEST_LGT))
+			{
+				SetNeedSendInfo();
+			}
+		}
+	}
+	MainView.HSplitTop(5.0f, nullptr, &MainView);
+
+	// Layout bottom controls and use remainder for skin selector
+	CUIRect QuickSearch, DatabaseButton, DirectoryButton, RefreshButton;
+	MainView.HSplitBottom(20.0f, &MainView, &QuickSearch);
+	MainView.HSplitBottom(5.0f, &MainView, nullptr);
+	QuickSearch.VSplitLeft(220.0f, &QuickSearch, &DatabaseButton);
+	DatabaseButton.VSplitLeft(10.0f, nullptr, &DatabaseButton);
+	DatabaseButton.VSplitLeft(150.0f, &DatabaseButton, &DirectoryButton);
+	DirectoryButton.VSplitRight(175.0f, nullptr, &DirectoryButton);
+	DirectoryButton.VSplitRight(25.0f, &DirectoryButton, &RefreshButton);
+	DirectoryButton.VSplitRight(10.0f, &DirectoryButton, nullptr);
+
+	// Skin selector
+	static CListBox s_ListBox;
+	std::vector<CSkins::CSkinListEntry> &vSkinList = SkinList.Skins();
+	int OldSelected = -1;
+	s_ListBox.DoStart(50.0f, vSkinList.size(), 4, 2, OldSelected, &MainView);
+	for(size_t i = 0; i < vSkinList.size(); ++i)
+	{
+		CSkins::CSkinListEntry &SkinListEntry = vSkinList[i];
+		const CSkins::CSkinContainer *pSkinContainer = vSkinList[i].SkinContainer();
+
+		if(!m_Dummy ? SkinListEntry.IsSelectedMain() : SkinListEntry.IsSelectedDummy())
+		{
+			OldSelected = i;
+			if(m_SkinListScrollToSelected)
+			{
+				s_ListBox.ScrollToSelected();
+				m_SkinListScrollToSelected = false;
+			}
+		}
+
+		const CListboxItem Item = s_ListBox.DoNextItem(SkinListEntry.ListItemId(), OldSelected >= 0 && (size_t)OldSelected == i);
+		if(!Item.m_Visible)
+		{
+			continue;
+		}
+
+		SkinListEntry.RequestLoad();
+		const CSkin *pSkin = pSkinContainer->State() == CSkins::CSkinContainer::EState::LOADED ? pSkinContainer->Skin().get() : pDefaultSkin;
+
+		Item.m_Rect.VSplitLeft(60.0f, &Button, &Label);
+
+		{
+			CTeeRenderInfo Info = OwnSkinInfo;
+			Info.Apply(pSkin);
+			vec2 OffsetToMid;
+			CRenderTools::GetRenderTeeOffsetToRenderedTee(CAnimState::GetIdle(), &Info, OffsetToMid);
+			const vec2 TeeRenderPos = vec2(Button.x + Button.w / 2.0f, Button.y + Button.h / 2 + OffsetToMid.y);
+			RenderTools()->RenderTee(CAnimState::GetIdle(), &Info, *pEmote, vec2(1.0f, 0.0f), TeeRenderPos);
+		}
+
+		{
+			SLabelProperties Props;
+			Props.m_MaxWidth = Label.w - 5.0f;
+			const auto &NameMatch = SkinListEntry.NameMatch();
+			if(NameMatch.has_value())
+			{
+				const auto [MatchStart, MatchLength] = NameMatch.value();
+				Props.m_vColorSplits.emplace_back(MatchStart, MatchLength, ColorRGBA(0.4f, 0.4f, 1.0f, 1.0f));
+			}
+			Ui()->DoLabel(&Label, pSkinContainer->Name(), 12.0f, TEXTALIGN_ML, Props);
+		}
+
+		if(g_Config.m_Debug)
+		{
+			Graphics()->TextureClear();
+			Graphics()->QuadsBegin();
+			Graphics()->SetColor(*pUseCustomColor ? color_cast<ColorRGBA>(ColorHSLA(*pColorBody).UnclampLighting(ColorHSLA::DARKEST_LGT)) : pSkin->m_BloodColor);
+			IGraphics::CQuadItem QuadItem(Label.x, Label.y, 12.0f, 12.0f);
+			Graphics()->QuadsDrawTL(&QuadItem, 1);
+			Graphics()->QuadsEnd();
+		}
+
+		// render skin favorite icon
+		{
+			CUIRect FavIcon;
+			Item.m_Rect.HSplitTop(20.0f, &FavIcon, nullptr);
+			FavIcon.VSplitRight(20.0f, nullptr, &FavIcon);
+			if(DoButton_Favorite(SkinListEntry.FavoriteButtonId(), SkinListEntry.ListItemId(), SkinListEntry.IsFavorite(), &FavIcon))
+			{
+				if(SkinListEntry.IsFavorite())
+				{
+					GameClient()->m_Skins.RemoveFavorite(pSkinContainer->Name());
+				}
+				else
+				{
+					GameClient()->m_Skins.AddFavorite(pSkinContainer->Name());
+				}
+			}
+		}
+
+		RenderSkinStatus(Item.m_Rect, pSkinContainer, SkinListEntry.ErrorTooltipId());
+	}
+
+	const int NewSelected = s_ListBox.DoEnd();
+	if(OldSelected != NewSelected)
+	{
+		str_copy(pSkinName, vSkinList[NewSelected].SkinContainer()->Name(), SkinNameSize);
+		SkinList.ForceRefresh();
+		SetNeedSendInfo();
+	}
+
+	static CLineInput s_SkinFilterInput(g_Config.m_ClSkinFilterString, sizeof(g_Config.m_ClSkinFilterString));
+	if(SkinList.UnfilteredCount() > 0 && vSkinList.empty())
+	{
+		CUIRect FilterLabel, ResetButton;
+		MainView.HMargin((MainView.h - (16.0f + 18.0f + 8.0f)) / 2.0f, &FilterLabel);
+		FilterLabel.HSplitTop(16.0f, &FilterLabel, &ResetButton);
+		ResetButton.HSplitTop(8.0f, nullptr, &ResetButton);
+		ResetButton.VMargin((ResetButton.w - 200.0f) / 2.0f, &ResetButton);
+		Ui()->DoLabel(&FilterLabel, Localize("No skins match your filter criteria"), 16.0f, TEXTALIGN_MC);
+		static CButtonContainer s_ResetButton;
+		if(DoButton_Menu(&s_ResetButton, Localize("Reset filter"), 0, &ResetButton))
+		{
+			s_SkinFilterInput.Clear();
+			SkinList.ForceRefresh();
+		}
+	}
+
+	if(Ui()->DoEditBox_Search(&s_SkinFilterInput, &QuickSearch, 14.0f, !Ui()->IsPopupOpen() && !GameClient()->m_GameConsole.IsActive()))
+	{
+		SkinList.ForceRefresh();
+	}
+
+	static CButtonContainer s_SkinDatabaseButton;
+	if(DoButton_Menu(&s_SkinDatabaseButton, Localize("Skin Database"), 0, &DatabaseButton))
+	{
+		Client()->ViewLink("https://ddnet.org/skins/");
+	}
+
+	static CButtonContainer s_DirectoryButton;
+	if(DoButton_Menu(&s_DirectoryButton, Localize("Skins directory"), 0, &DirectoryButton))
+	{
+		Storage()->GetCompletePath(IStorage::TYPE_SAVE, "skins", aBuf, sizeof(aBuf));
+		Storage()->CreateFolder("skins", IStorage::TYPE_SAVE);
+		Client()->ViewFile(aBuf);
+	}
+	GameClient()->m_Tooltips.DoToolTip(&s_DirectoryButton, &DirectoryButton, Localize("Open the directory to add custom skins"));
+
+	TextRender()->SetFontPreset(EFontPreset::ICON_FONT);
+	TextRender()->SetRenderFlags(ETextRenderFlags::TEXT_RENDER_FLAG_ONLY_ADVANCE_WIDTH | ETextRenderFlags::TEXT_RENDER_FLAG_NO_X_BEARING | ETextRenderFlags::TEXT_RENDER_FLAG_NO_Y_BEARING | ETextRenderFlags::TEXT_RENDER_FLAG_NO_PIXEL_ALIGNMENT | ETextRenderFlags::TEXT_RENDER_FLAG_NO_OVERSIZE);
+	static CButtonContainer s_SkinRefreshButton;
+	if(DoButton_Menu(&s_SkinRefreshButton, FontIcon::ARROW_ROTATE_RIGHT, 0, &RefreshButton) || Input()->KeyPress(KEY_F5) || (Input()->KeyPress(KEY_R) && Input()->ModifierIsPressed()))
+	{
+		ShouldRefresh = true;
+	}
+	TextRender()->SetRenderFlags(0);
+	TextRender()->SetFontPreset(EFontPreset::DEFAULT_FONT);
+
+	if(ShouldRefresh)
+	{
+		GameClient()->RefreshSkins(CSkinDescriptor::FLAG_SIX);
+	}
+}
